+++ examples/sending-mime-email-with-freemarker-example/src/main/java/com/test/MimeEmailWithFreemarkerApplication.java
@@ -12,6 +12,7 @@ import org.springframework.context.support.AbstractApplicationContext;
 
 import javax.annotation.PostConstruct;
 import java.io.UnsupportedEncodingException;
+import java.net.URISyntaxException;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
@@ -34,7 +35,7 @@ public class MimeEmailWithFreemarkerApplication implements ApplicationContextAwa
     }
 
     @PostConstruct
-    public void sendEmail() throws UnsupportedEncodingException, InterruptedException, CannotSendEmailException {
+    public void sendEmail() throws UnsupportedEncodingException, InterruptedException, CannotSendEmailException, URISyntaxException {
         testService.sendMimeEmailWithFreemarker();
 
         close();