+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Using Spring Boot _1.5.3.RELEASE_
+- Using Pebble _2.4.0.RELEASE_
 
 
 ## [0.6.2] -  2017-05-02